Implementando a consulta dos periodos na requisição.

+++ app/Http/Controllers/TarifaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Tarifa;
+use DB;
 
 class TarifaController extends Controller
 {
@@ -14,7 +15,15 @@ class TarifaController extends Controller
      */
     public function index()
     {
-        echo Tarifa::get();
+        $ini = $_REQUEST['data_ini'];
+        $fim = $_REQUEST['data_fim'];
+
+        if($ini) {
+            $dados = DB::select('select data_ini,data_fim,preco,estoque,disponivel from tarifas where data_ini = ? and data_fim = ?', [$ini,$fim]);
+        } else {
+            $dados = Tarifa::get();
+        }
+        return response()->json($dados);
     }
 
     /**
@@ -25,7 +34,12 @@ class TarifaController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        $resultado = Tarifa::create($request->all());
+        if($resultado) {
+            echo 'Tarifa cadastrada com sucesso!';
+        } else {
+            echo 'Erro ao cadastrar a tarifa!';
+        }
     }
 
     /**
@@ -36,7 +50,15 @@ class TarifaController extends Controller
      */
     public function show($id)
     {
-        echo Tarifa::find($id);
+        if($id) {
+            $dados = Tarifa::find($id);
+        }
+
+        if($dados) {
+            return response()->json($dados);
+        } else {
+            echo "Nenhuma tarifa encontrada.";
+        }
     }
 
     /**
@@ -48,7 +70,7 @@ class TarifaController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        dd($request);
     }
 
     /**
@@ -59,7 +81,12 @@ class TarifaController extends Controller
      */
     public function destroy($id)
     {
-        $resultado = Tarifa::find($id)->delete();
-        echo $resultado ? 'Tarifa removida com sucesso' : 'Não foi possível excluír a tarifa';
+        if(Tarifa::find($id)) {
+            $resultado = Tarifa::find($id)->delete();
+            echo $resultado ? 'Tarifa removida com sucesso' : 'Não foi possível excluír a tarifa';
+        } else {
+            echo 'Não foi possível excluír a tarifa';
+        }
     }
+
 }
